Updated inputs to match mvlearn and scikit-learn compatible
