+++ cca_zoo/model_selection/__init__.py
@@ -1,0 +1,1 @@
+from ._search import MVGridSearchCV
